feat(ast): Adding sizeof grammar rules. Refactoring `Factor` into all the constituent rules it was doing all at once and putting them into a private module.
