+++ src/ast/mod.rs
@@ -2,7 +2,8 @@ pub mod constants;
 pub mod declaration;
 pub mod declarators;
 pub mod exprs;
-pub mod factor;
+mod grammar;
+use grammar::*;
 pub mod operators;
 pub mod statements;
 mod token_stream;
@@ -20,7 +21,6 @@ pub use constants::Constant;
 pub use declaration::{Declaration, FunDecl, VarDecl};
 pub use declarators::{AbstractDeclarator, Declarator};
 pub use exprs::Expr;
-pub use factor::Factor;
 pub use operators::{BinaryOp, UnaryOp};
 pub use statements::Stmt;
 pub use types::{BaseType, Type, TypeBuilder};
